+++ tests/PSBlueSky.tests.ps1
@@ -2,29 +2,29 @@ BeforeDiscovery {
     # Import the module manifest
     $module = $PSCommandPath.Replace('.tests.ps1', '.psd1')
     #set a global variable so I can use this in my tests
-    $global:modulePath = Join-Path -Path .. -ChildPath (Split-Path $module -Leaf)
-    $ModuleName = (Get-Item $global:modulePath).BaseName
-    If (Test-Path $global:modulePath) {
-        Import-Module $global:modulePath -Force
+    $global:ModulePath = Join-Path -Path .. -ChildPath (Split-Path $module -Leaf) -Resolve
+    $ModuleName = (Get-Item $global:ModulePath).BaseName
+    If (Test-Path $global:ModulePath) {
+        Import-Module $global:ModulePath -Force -PassThru
     }
     else {
-        Write-Warning "Can't find module at $global:modulePath"
+        Write-Warning "Can't find module at $global:ModulePath"
     }
 }
 
 Describe "Module $ModuleName" -Tag Module {
     BeforeAll {
         #only need to get the module once
-        $thisModule = Test-ModuleManifest -Path $global:modulePath
+        Write-Host "Testing module at $global:ModulePath" -ForegroundColor cyan
+        $global:tm = $thisModule = Test-ModuleManifest -Path $global:ModulePath
     }
     AfterAll {
-        Remove-Variable -Name modulePath -Scope Global
+        Remove-Variable -Name ModulePath -Scope Global
     }
 
     Context Manifest {
         It 'Should have a module manifest' {
-            $modulePath | Should -Exist
-            $thisModule.ModuleType | Should -Be 'Script'
+            $global:ModulePath | Should -Exist
         }
         It 'Should have a defined RootModule' {
             $thisModule.RootModule | Should -Be "$($thisModule.Name).psm1"
@@ -55,50 +55,208 @@ Describe "Module $ModuleName" -Tag Module {
         }
         It 'Should have a LicenseUri' {
             $thisModule.PrivateData.PSData.licenseUri | Should -Not -BeNullOrEmpty
-        } -pending
+        }
         It 'Should have a defined ProjectURI' {
             $thisModule.PrivateData.PSData.ProjectUri | Should -Not -BeNullOrEmpty
-        } -pending
+        }
         It 'Should have defined tags' {
             $thisModule.PrivateData.PSData.Tags.count | Should -BeGreaterThan 0
-        } -pending
+        }
 
     }
     Context 'Module Content' {
         It 'Should have a changelog file' {
-            "..\changelog.md" | Should -Exist
+            '..\changelog.md' | Should -Exist
         }
         It 'Should have a license file' {
-            "..\license.txt" | Should -Exist
+            '..\license.txt' | Should -Exist
         }
-        It 'Should have a README file'  {
-            "..\README.md" | Should -Exist
+        It 'Should have a README file' {
+            '..\README.md' | Should -Exist
         }
         It 'Should have a docs folder' {
-            "..\docs" | Should -Exist
+            '..\docs' | Should -Exist
         }
         It 'Should have a markdown file for every exported function' {
-            $functions =$thisModule.ExportedFunctions
+            $functions = $thisModule.ExportedFunctions
             $functions.GetEnumerator() | ForEach-Object {
-                $mdFile = "..\docs\$(.Key).md"
+                $mdFile = "..\docs\$($_.Key).md"
                 $mdFile | Should -Exist
             }
-        } -pending
+        }
         It 'Should have external help' {
-            "..\en-us"  | Should -Exist
-            "..\en-us\OSDetail-help.xml" | Should -Exist
-        } -pending
+            '..\en-us' | Should -Exist
+            '..\en-us\PSBluesky-help.xml' | Should -Exist
+        }
         It 'Should have a Pester test' {
             #this is probably silly since I'm using a Pester test.
-            ".\*tests.ps1" | Should -Exist
+            '.\*tests.ps1' | Should -Exist
         }
     }
 }
 
-InModuleScope $Module {
-    Describe FunctionToTest {
-        It "does something useful" {
-            $true | Should -Be $true
-            }
+InModuleScope $ModuleName {
+
+    Describe New-BskyPost {
+        It 'Should have help documentation' {
+        (Get-Help New-BskyPost).Description | Should -Not -BeNullOrEmpty
         }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name New-BskyPost).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { New-BskyPost } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyTimeline {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyTimeline).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyTimeline).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyTimeline } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskySession {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskySession).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskySession).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskySession } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyProfile {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyProfile).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyProfile).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyProfile } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyFollowing {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyFollowing).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyFollowing).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyFollowing } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyFollowers {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyFollowers).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyFollowers).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyFollowers } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyFeed {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyFeed).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyFeed).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyFeed } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Get-BskyAccessToken {
+        It 'Should have help documentation' {
+        (Get-Help Get-BskyAccessToken).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Get-BskyAccessToken).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Get-BskyAccessToken } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
+    Describe Add-BskyImage {
+        It 'Should have help documentation' {
+        (Get-Help Add-BskyImage).Description | Should -Not -BeNullOrEmpty
+        }
+        It 'Should have a defined output type' {
+        (Get-Command -CommandType function -Name Add-BskyImage).OutputType | Should -Not -BeNullOrEmpty
+        }
+        It 'Should run without error' {
+            <#
+        mock and set mandatory parameters as needed
+        this test is marked as pending since it
+        most likely needs to be refined
+        #>
+            { Add-BskyImage } | Should -Not -Throw
+        } -Pending
+        #insert additional command-specific tests
+
+    } -Tag function
 }